finagle-core: Fix flaky test
Summary: Problem

DefaultTracingTest uses the same tracer for both client and server, but
the response events have been observed to be reordered from time to
time. This is likely due to recording those events via a `respond`
block, which doesn't guarantee an order with respect to the other
continuations attached to the Future.

Solution

Don't interleave client and server events in the same Tracer.

JIRA Issues: CSL-6294

Differential Revision: https://phabricator.twitter.biz/D162478