finagle-netty4: Propagate ChannelExceptions from Netty4PushChannelHandle.onClose
Summary: Problem

Some exceptions from the Netty channel may be of interest to the
application layer, such as SSLExceptions, but we don't propagate them at
all.

Solution

If the handle is closed due to an exception other than
ChannelClosedException, we propagate that through the failure channel of
the `onClose` promise.

JIRA Issues: CSL-6261

Differential Revision: https://phabricator.twitter.biz/D160475