finagle-toggle: Improve java compatibility
Summary: Problem

ToggleMap and Toggle found a bunch of java compatibility landmines.

1. traits that extend abstract classes don't propagate their inheritance
to java.
2. methods that return generics with a primitive subtype are compiled
as if they return an object subtype.  jdk8 disallowed casting from Gen<Object>
to Gen<SpecificType>.
3. PartialFunction[A, Boolean]#apply returns an Object instead of a boolean.

Solution

Add helpers to ToggleMap and Toggle to return a better type in java, and change
everything that extends ToggleMap to be an abstract class.  For ToggleMap.Proxy,
which is mixed in with other classes, we should instead change it to no longer
extend ToggleMap, and instead have a self-type which conforms to ToggleMap.

JIRA Issues: CSL-6325

Differential Revision: https://phabricator.twitter.biz/D164489