finagle-memcached: support partial success for the batch operations in the new PartitioningService based memcached client
Summary: Both the old and new memcached clients use com.twitter.finagle.memcached.ConnectedClient, which
marks the failed records in batch calls as failures and so the caller needs to retry only the
failed records. The old client wraps per partition requests with ConnectedClient so only the
records corresponding to the failed partition are marked as failed. The new client wraps the
entire request with ConnectedClient since PartitioningService works at a very low level. Also
PartitiongService fails the entire batch as soon as it receives failure from one partition. So
all records are retried leading to higher failure counts.

Solution:

Now PartitionService waits for all futures to complete and returns both successful responses and
a map from keys to exceptions. MemcachedPartitioningService layer aggregates them in a new
Response type "ValuesAndErrors" and returns it upstream and the ConnectedClient handles the
partial success as expected.

Result:

When the memcache shards are slow the number of retries are now reduced and also failure counts
are drastically reduced.

JIRA Issues: FES-4000

Differential Revision: https://phabricator.twitter.biz/D161249