finagle-core: Fix flaky DefaultTracingTest
Summary: Problem

DefaultTracingTest is flaky, potentially due to a race on the server
between sending the response and adding the tracing annotation. We
only await the response on the client side, so it's possible that the server
adds the annotation after the client has already received it.

Solution

Use an `eventually` block to allow the racy behavior to resolve.

JIRA Issues: CSL-6347

Differential Revision: https://phabricator.twitter.biz/D165777