finagle-http: Add final modifier to string vals in Fields
Problem
----------
It would be less tedious and error-prone to be able to use the string constants in
finagle http `Fields` with Java's annotation. Currently, we would get a compiler
error if we try to do that. For instance with Finatra:

```
case class CustomRequest(
  @Header(Fields.AcceptLanguage) acceptLanguage: Option[String],
  ...
)
```

`annotation argument needs to be a constant; found: com.twitter.finagle.http.Fields.AcceptLanguage`

Solution
----------
We can reference the string constant in the annotation by adding the `final`
modifier.

Differential Revision: https://phabricator.twitter.biz/D166843