finagle-netty4: Supply Correct Address to SslServerSessionVerifier
Problem / Solution

Currently `SslServerSessionVerifier` always receives an address of
`Address.failing`. This is due to incorrect logic in checking
that the channel's remote address is an `InetSocketAddress`.
It is accidentally checking whether the `Channel` itself is an
instance of `InetSocketAddress`, which is never the case. Change
the logic to test whether the `Channel.remoteAddress` is an instance
of `InetSocketAddress` instead. Then the address is properly supplied
to the `SslServerSessionVerifier`.

JIRA Issues: CSL-6262

Differential Revision: https://phabricator.twitter.biz/D168334