finagle-core: Reduce lock contention in Closables
Problem

We have to register and deregister a closable in Closables every time we receive
a new request and then respond to it in h2.  High throughput servers have seen
lock contention from this.

Solution

Use a ReadWriteLock instead of an intrinsic lock, and switch to using a CHM
instead of a HashSet.

JIRA Issues: CSL-5971

Differential Revision: https://phabricator.twitter.biz/D171480