finagle-core: Add toString to c.t.finagle.{Filter, Service}
Summary: Problem

We'd like increased visibility into the make up of composed
Filters (and the terminal Service or ServiceFactory) in a
Filter chain which helps understand what is going on in a
myriad of situations including when step-through debugging.

Solution

Add `toString` implementations for `c.t.finagle.Filter` and
`c.t.finagle.Service` and properly manipulate the toString
implementation during `andThen` composition.

Result

A better default `toString` implementation for Service,
ServiceFactory and ServiceProxy, a more useful `toString`
for a Filter composition and for a composed Service (a Filter
chain with a terminal Service or ServiceFactory).

JIRA Issues: CSL-5355

Differential Revision: https://phabricator.twitter.biz/D172526