finagle-http: test that empty buf doesn't close stream prematurely
HTTP streams rely on chunked transfer encoding, which encodes the end of
stream as an empty chunk. Our design intent is to shield users from the
underlying encoding, such that writing an empty Buf to an HTTP stream
must not translate to an empty chunk (prematurely signaling EOS) - this
would be a very serious bug.

These tests show that no such bug is present.

JIRA Issues: CSL-2742

Differential Revision: https://phabricator.twitter.biz/D176670