finagle-base-http: Remove semicolon from N4CC's set of wrappable chars
Problem

The comment above `ShouldWrapCharsBitSet` suggests that semicolons are legal
according to Netty 4's encoder and decoder, but they are not.

Solution

Remove the semicolon character from the bit set.

Result

Now the bit set is correctly aligned with Netty 4's strict validation rules and
the comment is correct. There is no difference in behavior since validation
happens after unwrapping: a semicolon in a cookie value will cause the
validation to fail regardless of whether the value is wrapped.

JIRA Issues: CSL-6592

Differential Revision: https://phabricator.twitter.biz/D184708