finagle-benchmark: Add failure mode to LB simulation
Problem

The load balancer simulation does not account for node failures.

Solution

Similar to the LatencyProfile, a FailureProfile is defined to tell
whether next incoming response should be failed or fulfilled.
The failure profile can be loaded from a file using experimental data
or instantiated and given durations. The ServerNode uses the profile
to determine how to respond to a request.

Result

Additional available flag when running the simulation that adds an
unhealthy backend node to the server set that fails for 33% of the simulation.

JIRA Issues: CSL-6584

Differential Revision: https://phabricator.twitter.biz/D183789