finagle/finagle-core: Accept double in Ema.update
Problem:

The value for the Ema can be reset or initialized to a specific value through
the first update call. Internally, the ema value is of the Double data type but
the update function accepts a Long. This prevents updating, initializing, or
resetting the ema to a decimal value even though it is stored as a decimal
internally.

Solution:

Change the api for updating the ema to accept a Double instead of a Long.

Differential Revision: https://phabricator.twitter.biz/D192938