finagle-serversets: Consider ip address when sorting Addresses
Problem

If a serverset comes back with all the same metadata which can happen
due to announcement bugs we don't have anything to sort against.

Solution

In the case where the shard id metadata is the same for both addresses
we fall back to sorting based on ip address and port. Since these are
unique, there is a unique ordering for the serverset.

JIRA Issues: CSL-6675

Differential Revision: https://phabricator.twitter.biz/D192203