finagle-mysql: Correctly convert java.sql.{Date,Timestamp} to Parameter
Problem

The java.sql.Date and java.sql.Timestamp implicits are ignored and the
java.util.Date conversion is used instead, because of the contravariance of the
CanBeParameter type parameter and the fact that java.sql.Date and
java.sql.Timestamp extend java.util.Date.

Solution

Use only one implicit (for the supertype) and pattern match within it to create
the correct Parameter.

JIRA Issues: CSL-6021

TBR=true

Differential Revision: https://phabricator.twitter.biz/D195351