finagle-core: introduce Stack.NamedTransformer
Problem

There isn't a good way to customize a Finagle server for a specific operating
environment. One approach we've taken with server admission control is to
insert a placeholder Stack module (ServerAdmissionControl) into which we can
register user-defined admission control filters.  This works fine but we need a
way to do this more generally. There should be a standard mechanism for
transforming the default shape of the Stack.

Solution

NamedTransformer is a standard mechanism for transforming the default shape of
the Stack. It is a Stack.Transformer with a name.  Registration and retrieval
of transformers from global state is managed by StackServer.DefaultTransformer.
The transformers will run at materialization time for Finagle servers, allowing
users to mutate a Stack in a consistent way.

JIRA Issues: CSL-6447

Differential Revision: https://phabricator.twitter.biz/D193922