finagle-core: introduce Stack.map; protect Stack.transform
Problem

Stack.transform is too powerful. We want its ability to modify each node
in the Stack without the ability to grow/shrink the Stack, and also
without the ability to modify params.

Solution

To do this, we reinstate the protected status of Stack.transform and
offer a safer alternative: Stack.map. Stack.map is similar to the map
operation over lists: it applies a function to each node in the stack,
creating a new stack, but it does so without growing/shrinking the
stack, and the function can't access params.

JIRA Issues: CSL-6757

Differential Revision: https://phabricator.twitter.biz/D195938