finagle-mysql: rollback off the request path
Problem

It's possible for a rollback from `RollbackFactory` to
take a significant amount of time and, thus, shouldn't be on
the request path.

Solution

Modify `RollbackFactory` to issue a rollback when we put
a connection back into the pool rather than when we check one out.

Result

As written, the additional work incurred for the rollback may
result in less  throughput from the connection pool, however, this
gives users a clear knob for addressing the issue (i.e. via configuring
the pool to offer more available connections).

JIRA Issues: CSL-6760

Differential Revision: https://phabricator.twitter.biz/D196673