finagle-http2: Avoid a race condition when streaming
Problem

When streaming over HTTP2, there's a race between the StreamTransport being
closed and another request coming through. This puts the StreamTransportFactory
in a bad state and the connection must be reestablished. This is particularly
easy to encounter when a reader is discarded while another request is in flight.

Solution

Check a volatile marker when looking at status that indicates whether or not a
stream has been closed. Set the marker before bouncing the rest of the work
through the executor.

JIRA Issues: CSL-6542

Differential Revision: https://phabricator.twitter.biz/D198198