finagle-mux: Use encapsulation for application headers in mux protocol
Problem

We currently mix application headers with traditional mux
broadcast context entries but don't have a strategy to separate
them on the peer. This leads to mixing and due to the need to
allow multiple application values for the same key, this can lead
to amplification of context entries.

Solution

Encapsulate application headers into a single broadcast context
entry. This gives us a mechanism for distinguishing the application
and broadcast context entries and allowing the different
header semantics.

JIRA Issues: CSL-6662

Differential Revision: https://phabricator.twitter.biz/D195974