finagle-core: Introduce `SslTransportContext`
Problem / Solution

`Transport.peerCertificate` is one small piece of information regarding the security of
the underlying transport for each request. The presence or lack of peer certificate is
not enough in many cases to determine the security of the connection. Let's introduce
new types for migrating towards a more feature-rich context for security information.

JIRA Issues: CSL-6323

Differential Revision: https://phabricator.twitter.biz/D203032