finagle-http: Make assertions more informative
Problem

We've seen some failures in Travis that only say that a Future was
defined but we don't know if it failed or succeeded, so it's not very
informative.

Solution

Check the structure of the result so that we can see if it failed,
succeeded, or wasn't yet defined.

JIRA Issues: CSL-6860

Differential Revision: https://phabricator.twitter.biz/D204799