finagle-http: Update HttpTransportTest to Test Transport.status
Problem / Solution

Methods such as 'status' on `Transport` are deprecated in favor of the
`TransportContext` versions. However, there are certain subtle pieces
of functionality incorporated into existing Transports, which are not
being tested and make migration more difficult. `HttpTransport.status`
is one of those. Let's add a portion of a test that will fail if that
functionality is not adequately migrated.

Differential Revision: https://phabricator.twitter.biz/D205285