finagle-netty4: allow sockets to be configured with the SO_REUSEPORT option
Problem

Would like to be able to build servers that open sockets with the
SO_REUSEPORT option on Linux. SO_REUSEPORT allows multiple processes
to bind and accept connections from the same address and port.

Our particular use-case involves being able to support near
zero-downtime restarts with finagle servers. We'd like to be able to
start a new, simultaneously-running finagle process, listen on the
same interface and port and then gracefully shutdown the old PID when
the new process is healthy.

Solution

Support for SO_REUSEPORT is available when using the native-epoll
transport in netty4 (netty/netty@d1d8a6b). Allow this option to be
configured using Transport.Options.

Signed-off-by: Kevin Oliver <koliver@twitter.com>

JIRA Issues: CSL-6895

Differential Revision: https://phabricator.twitter.biz/D205535