finagle-core: Use PendingRequestFilter or RequestSemaphoreFilter based on maxWaiters
Problem

RequestSemaphoreFilter is configured with a semaphore that's
commonly constructed with a maxWaiters of zero (e.g, configuring
maxConcurrentRequests on a server creates an AsyncSemaphore
with a maxWaiters of zero). In this case, we pay the unnecessary cost
of lock synchronization and submitting a future callback to the scheduler.

Solution

If maxWaiters is zero, create a PendingRequestFilter, which performs better.
Otherwise, create a RequestSemaphoreFilter.

Differential Revision: https://phabricator.twitter.biz/D194207