finagle-core: make interrupts configurable in SingletonPool
Problem

SingletonPool guards an underlying shared resource. However, it allows
interrupts to propagate directly to the underlying service acquisition
requests which can cause one interrupt (e.g. a timeout) to kill the
acquisition for others waiting on the same resource.

Solution

Instead, this behavior should be configurable and up to the discretion
of the protocol implementation to determine. We introduce a parameter
for this which is passed into the constructor for `SingletonPool`.

Differential Revision: https://phabricator.twitter.biz/D205433