finagle-netty4: Change `ChannelTransport` to Use `ChannelTransportContext`
Problem

`Transport`s which use `TransportContext`s which delegate
back to the original `Transport` prevent deprecated methods
from being removed from `Transport` entirely. Finagle-Netty4's
`ChannelTransport` is one of those.

Solution

Rename `Netty4Context` to `ChannelTransportContext` and change
the methods on it to stop delegating back to a `Transport`.

Result

One more `Transport` is using a proper `TransportContext` and is
no longer blocking the removal of deprecated methods on `Transport`.

Differential Revision: https://phabricator.twitter.biz/D205794