finagle-core, finagle-mux: Introduce some conventions for using the single executor pattern
Problem

The serial executor concurrency pattern is great for managing
concurrency but it can be tough to know what state is safe to modify
when.

Solution

Introduce some conventions for using the serial executor. First, methods
that are intended to be called from within the serial executor are
prefixed with the word `handle`. Second, fields that should only be used
from within the serial executor is prefixed with `h_`.

Example:
`def handleMutateSomeState(): Unit = { h_counter += 1 }`

Add a README.md

JIRA Issues: CSL-6784

Differential Revision: https://phabricator.twitter.biz/D203680