finagle-http2: Fixes race condition on h2c upgrade
Problem

We're seeing a race condition in travisci where we close the reftransport and
then try to update it.

Solution

Simplify the model so that it's OK to close the reftransport and continue using
it, but we make sure to call close on every subsequent transport that it's
updated with.

JIRA Issues: CSL-6925

Differential Revision: https://phabricator.twitter.biz/D209440