finagle-netty4: Use FastThreadLocalThreads for finagle worker pools
Problem

The finagle-netty4 implementation overrides the default thread
factory supplied by netty to its event loop groups.  By doing so,
it no longer creates special `FastThreadLocalThread`s that netty
is optimized for, but instead normal threads.

Solution

Use the netty `DefaultThreadFactory` when creating the executor for
the netty event loop groups.  This will create `FastThreadLocalThread`s,
as was as continue to provide the other functionality the finagle
customizations provided.

Differential Revision: https://phabricator.twitter.biz/D207981