finagle-http: Use singleton pool in some streaming tests
Problem

Some streaming tests assume that a second request issued while a first request
is outstanding will be in a pending state. In non-multiplxed HTTP/1x, this was
often the case. However, with multiplexing, this becomes flaky.

Solution

For testing that sort of behavior, use a pool size of 1.

JIRA Issues: CSL-6939

Differential Revision: https://phabricator.twitter.biz/D212297