finagle-serversets: initialize zkHealth to Unknown
Problem

On start up, the zkHealth reports "Healthy" when the client is not able
to connect to zookeeper. On debugging, I noticed that the ClientHealth
was being initialized to Healthy.

Solution

The ClientHealth should be initialized to Unknown, and it will become
Healthy only when it connects to zookeeper (which internally means the
session changing to SyncConnected)

Result

If the client isn't able to connect to zookeeper on startup, it will
report its health as Unknown.

Signed-off-by: Kevin Oliver <koliver@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D213768