finagle-base-http: Dedupe HTTP 1.1 header validation
Problem

When we convert HTTP 1.1 headers from Netty to Finagle, we validate the headers
twice: once in Netty and once in Finagle.

Solution

Create an API for adding headers to the HeaderMap without validating them and
use this method in the Netty -> Finagle bijection.

Result

We dedupe this validation and set ourselves up for deduping other validations
that occur in Finagle's HTTP projects.

JIRA Issues: CSL-6990

Differential Revision: https://phabricator.twitter.biz/D217035