finagle-http2: alpn/npn wait for singleton session to resolve for concurrent requests
Problem

For ALPN or NPN based HTTP/2 session negotiation it's nice to just wait
for the result of negotiation for racing second requests instead of falling
back to HTTP/1.x.

Solution

For the ALPN or NPN pathway we wait while for HTTP/1.x we make
temporary HTTP/1.x sessions.

JIRA Issues: CSL-7005

Differential Revision: https://phabricator.twitter.biz/D217855