finagle/travis: Introduce SKIP_FLAKY_TRAVIS env var
Problem

Some tests are only flaky in Travis CI. It's useful to be able to mark them
as such without using our usual SKIP_FLAKY marker which causes the tests to
be skipped in submit queue as well.

Solution

Check a new environment variable, SKIP_FLAKY_TRAVIS, instead and set it for
travis CI runs.

Differential Revision: https://phabricator.twitter.biz/D220184