finagle-netty3: Rebalance Deprecated Transport and TransportContext Methods
Problem

It has been decided that the proper balance of deprecated methods on
`Transport` is to only move half of them to `TransportContext`.

Solution

Remove the deprecation warnings on some methods on `Transport`, and add them to
those methods on `TransportContext`. Also, start the migration process by
fixing finagle-netty3.

JIRA Issues: CSL-7050

Differential Revision: https://phabricator.twitter.biz/D221528