finagle-core: improve precision in Ring.range
Problem

The range method in [[Ring]] was too coarse grained when
detecting the boundary cases. This causes issues when
computing a `width` that is near 1.0, for example.

Solution

Compute the boundary cases based on the start and end indices
rather than absolute values of `width` which are hard to compare
at arbitrary precision.

JIRA Issues: CSL-7038

Differential Revision: https://phabricator.twitter.biz/D221127