finagle-http2: Evict dead sessions when they are encountered in Http2NegotiatingTransporter
Problem

When a HTTP/2 client session closes, either due to draining or a problem,
we will encounter it on the next iteration through the Http2Transporter and
yield a DeadTransporter. This will go against the clients retry budget but it's
a very normal thing to have happen and thus shouldn't.

Solution

If the Http2Transporter encounters a dead ClientSession it will evict it and
try again. We don't encounter a loop since the re-initialization of the ClientSession
will also yield the Transport to use for the initiating request, thus terminating the
loop.

JIRA Issues: CSL-7037

Differential Revision: https://phabricator.twitter.biz/D220597