finagle-http2: Change the status computation of Http2NegotiatingTransporter
Problem

The Http2NegotiatingTransporter will bubble up the status of a dead
ClientSession as Status.Closed and cause higher levels to avoid this
endpoint.

Solution

Consider a cached session in the closed state to really be a Status.Open
since it will be evicted and we're happy to establish a new session.

Make status logic easier to read

JIRA Issues: CSL-7088

Differential Revision: https://phabricator.twitter.biz/D226416