finagle-core: Introduce Filter.OneTime
Problem

Filter.TypeAgnostic is actually a factory that has a toFilter[Req, Rep] method
which is called to create a new filter of those types. Often, a TypeAgnosic
filter only has toFilter called once per instance. We want to encourage folks to
use Filter.TypeAgnostic filters in Finatra & elsewhere, but it's not as easy as
it could be to create them.

Solution

Introduce a type, Filter.OneTime, that removes a lot of boiler plate when
creating a TypeAgnostic filter.

Result

Instead of this:

  class SaveWhitespaceFilter extends Filter.TypeAgnostic {
    def toFilter[Req, Rep]: Filter[Req, Rep, Req, Rep] =
      new Filter[Req, Rep, Req, Rep] {
        def apply(req: Req, svc: Service[Req, Rep]): Future[Req, Rep] = {
          ... implementation ...
        }
      }
  }

A TypeAgnostic filter can be created like this:

  class SaveWhitespaceFilter extends Filter.OneTime {
    def apply[Req, Rep](req: Req, svc: Service[Req, Rep]): Future[Rep] = {
      ... implementation ...
    }
  }

Differential Revision: https://phabricator.twitter.biz/D228345