finagle-redis: Fix flaky integration tests
Problem

We have a race condition in PubSubClientIntegrationSuite2 where on a
subscription update we satisfy the pending Promise and then update the
message count map. In the assert, we only await the Promise resolution
before checking the map, meaning that if the update pauses after setting
the Promise the asserting thread exits it's `Await.result` successfully
only to find out that the message count is still 0.

Solution

Make the updates to the two different collections atomic through
explicit synchronization.

JIRA Issues: CSL-7147

Differential Revision: https://phabricator.twitter.biz/D230965