finagle-mux: Cleanup gauges when the session closes
Problem

We have some gauges that end up dangling and are not cleaned up when the
session closes. Since mux sessions are commonly long lived the
application typically needs to wait for an old-gen GC cycle to remove
them, and those GC cycles are hopefully infrequent.

Solution

Explicitly remove the gauges when the session closes.

JIRA Issues: CSL-7148

Differential Revision: https://phabricator.twitter.biz/D231534