finagle-mysql: introduce session
Problem

Many MySQL operations require keeping session state, this recently came up in
the context of advisory locks. As it turns out, there are several uses for this
functionality. The current workaround is to do all session work within a
transaction but this has the effect of wrapping all session operations in a
transaction, which may not be the desired behavior.  Since this is a common
pattern in MySQL we should expose the ability for users to "check-out" a session
for the duration of some series of operations.

Solution

Provide a method, `session` that allows a user to reserve a client connection so
that they may perform operations that require client/server state. This also
adds the ability for the user to `discard` a session if they have determined that
an operation has put the session in an unrecoverable state.

JIRA Issues: CSL-7018

TBR=true

Differential Revision: https://phabricator.twitter.biz/D219322