finagle-mysql: Make Integration Tests More Resilient
Problem

One finagle-mysql integration test can fail if the
integration tests are run as a non-root user, which
they should be to mimic production environments. Another
can fail if the default character set of the MySQL server
is not set to a particular value.

Solution

Change the integration tests to work regardless of whether
a root or non-root user is used and when the default MySQL
character set is used.

JIRA Issues: CSL-7174

Differential Revision: https://phabricator.twitter.biz/D233142