finagle-http2: Minor cleanup
Problem

During the normal churn a few things get out of whack such as the name
of a ChannelHandler and we were clobbering a setting that we shouldn't
have been and the behavior of the `Http2ClientEventMapper` didn't capture
the event in the right exception pathway.

Solution

Clean things up. The `Http2ClientEventMapper` now transforms
`HeaderListSizeException`s in the write promise instead of the `exceptionCaught`
pathway since this is where they now surface in `Http2MultiplexCodec`.

Differential Revision: https://phabricator.twitter.biz/D234115