finagle-mysql: Sequence the singleton session in the Session
Problem

When getting a singleton service for use with queries that require the
same session, like transactions, we internally store the Future[Service]
from the underlying call meaning that we return a client eagerly, but
we don't have a backing session.

Solution

We can do better by waiting until we have a backing session to return
session client. In the failure to acquire case we avoid the overhead of
making queries against the pinned session client.

Differential Revision: https://phabricator.twitter.biz/D234619