finagle-http2: Even more graceful draining of H2 server sessions
Problem

When we drain a server we immediately send a GOAWAY with the last
observed stream id. This tells the client that all streams after that id
were not processed and are safe to retry. However, if the client is
racing a request with this GOAWAY that it cannot retry it's still a
lost request.

Solution

Instead of sending the GOAWAY with the last observed stream id we send
it with Int.MaxValue. This lets racing streams be processed. If they
don't terminate within the deadline we force the session closed to
prevent resource leaks.

JIRA Issues: CSL-6803

Differential Revision: https://phabricator.twitter.biz/D206683