finagle-core: Handle empty address sets in TrafficDistributor
Problem

Finagle's TrafficDistributor is used to build endpoints with their associated
weight classes. This endpoint-to-weight class grouping is used to create load
balancers used to balance requests for a particular endpoint stack. A
TrafficDistributor's load balancer are tracked in the global BalancerRegistry
every time there is an address set change for a particular service. The issue is
that TrafficDistributor has two places in its code that creates a load balancer
and one of these balancers never gets destroyed when a service is closed. This
causes a leak in BalancerRegistry This scenario is evident when a service's
address set alternates periodically between an empty address set to a valid
populated address set.

Solution

This PR removes code that extraneously creates an empty load balancer and adds
it to the partition method in TrafficDistributor. This causes TrafficDistributor
to manage address state in one area and avoids the object leak.

Signed-off-by: Jordan Parker <jparker@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D244335