finagle-redis: Add support for Stream APIs
Problem

Redis 5.0 was released recently and brought with it the new Stream
API. The finagle-redis client should support these new APIs.

Solution

Added support for the Stream API in the finagle-redis client.

Result

Users of the finagle-redis library will be able to access the new Stream
API which is now generally available on Redis >=5.0.

Signed-off-by: Jordan Parker <jparker@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D244329