finagle-mysql: Remove `Handshake` from `ClientDispatcher` API
Problem

When TLS is added to finagle-mysql, the `Handshake` will no longer
occur as part of the `ClientDispatcher` and will instead happen as
part of an eventual `MysqlTransporter`. Currently, `Handshake` is
part of the `ClientDispatcher` construction API. This prevents it
from being moved and eventually removed.

Solution

Change the `ClientDispatcher` to take a generic set of `Stack.Params`,
and make the `Handshake` an implementation detail.

JIRA Issues: CSL-7284

Differential Revision: https://phabricator.twitter.biz/D244357