util-core: Provide a way to listen for stream termination
Problem

It can be useful to know when a stream is closed for implementing many common
questions, like "how long has my stream taken to finish" or "did my stream
finish successfully?" but there is not currently a good way to tell.

Solution

Provide Reader#onClose which will be satisfied after the stream is finished,
discarded, or failed.

JIRA Issues: CSL-6813

TBR=true

Differential Revision: https://phabricator.twitter.biz/D236311