finagle-zipkin-core: Hold completed spans in DeadlineSpanMap for a short time
Problem

Zipkin spans are stored in a local cache (the DeadlineSpanMap) until the span is
deemed complete by the generation of either a ClientRecv or ServerSend
annotation on the client or server, respectively. Once marked complete, the span
is removed from the cache and logged to the collector. Any annotations generated
subsequently for a completed span will eventually be sent to the collector after
a configurable timeout period (currently 120s), but will be orphaned from the
rest of the trace and possibly lost.

Solution

The hard timeout is necessary to avoid spans sitting in the cache forever, while
eager span completion is important for timely storing of traces in the backend.
However, the hard-coding of ClientRecv/ServerSend as being the "final"
annotations at each end of the RPC causes subsequently generated annotations to
be orphaned. This change instead uses those annotations as hints that the span
will complete "soon" and waits a short (configurable) amount of time for any
subsequent annotations that may be generated by the same process before closing
the span.

Result

Spans are delayed a short amount of time (500ms by default) at the host. Fewer
span fragments, comprising one or two annotations and orphaned from the rest of
the span, are sent to the collector.

JIRA Issues: CSL-7253

Differential Revision: https://phabricator.twitter.biz/D244144