finagle-zipkin-core: Eagerly log spans after hold timer expires
Problem

The hold timer is used to keep spans on the host for a short amount
of time (by default 500ms) following generation of a ClientRecv or
ServerSend annotation. After the timer expires, spans are considered
ready to send to the Zipkin collector, but the span is not logged until
either the ttl expires or the span is updated again. However, spans
are typically not updated following expiry of the hold timer and so
we see undesirable delays in logging spans to the collector.

Solution

Log the span to the collector as soon as the hold timer expires.

Result

In most cases, the delay writing traces to persistent storage is
determined by the short hold timer rather than the long ttl
(default 120s).

One side effect is that it is possible the hold timer and the ttl
timer will race to log the span, which may result in duplicate
annotations. Zipkin already has logic to handle this situation.

JIRA Issues: CSL-7253

Differential Revision: https://phabricator.twitter.biz/D249641