finagle-http2: Even better shutdown in H2Filter
Problem

If the close deadline had already expired we don't attempt the smart
graceful shutdown, and also don't set the Http2ConnectionHandlers
graceful shutdown time, which by default is set to 30 seconds. If we
don't send a GOAWAY ourselves, as in the case of the deadline already
being expired, we end up using the default.

Solution

If we don't send a GOAWAY ourselves then set the deadline.

JIRA Issues: CSL-7372

Differential Revision: https://phabricator.twitter.biz/D250492