finagle-core: allow the ability to set a global trace service name
Problem

It's useful to derive a service name for a trace from the environment.
For example, using a process coordinate can be more descriptive than
the user specified label.

Solution

Expose an API which allows us to set the label globally and prefer
it when it is available.

JIRA Issues: CSL-7279

TBR=true

Differential Revision: https://phabricator.twitter.biz/D242655